Move eventCallback for progress
