+++ src/store/model.ts
@@ -502,6 +502,12 @@ export interface IJobEventData {
   status?: number;
 }
 
+export interface IProgressInfo {
+  title?: string;
+  info?: string;
+  progress?: number;
+}
+
 export interface IComputeJob {
   jobId: string;
   callback: (success: boolean) => void;
